refactor: break SQLAlchemy / cloud_storage circular-import clusters (wlvxh)
Resolves 15 CodeQL py/cyclic-import findings across two clusters.

Cluster A — cloud_storage factory cycle (8 findings):
- Split cloud_storage/base.py into types.py (ABC + dataclasses) and
  factory.py (get_adapter only). Delete base.py entirely — no shim,
  since a shim re-introduces the cycle at the import graph level.
- Concrete adapters (s3/gdrive/onedrive/dropbox) now import from
  cloud_storage.types. factory.py lazily imports concrete adapters.
  One-way edges only: types -> adapters -> factory.
- cloud_storage/__init__.py re-exports the public surface unchanged.
- Update external callers (publish_pipeline, routers/export, three
  test modules) to import from cloud_storage package root.

Cluster B — database <-> models cycle (7 findings):
- Extract Base = declarative_base() into new backend/db_base.py.
- database.py now imports Base from db_base and re-exports it, so
  existing `from database import Base` call sites (ffmpeg_builder,
  alembic/env, conftest, tests) keep working.
- models.py, export_models.py import Base from db_base instead of
  database. database's lazy `from models import ...` inside init_db()
  is now one-way — no cycle.
- normalization_migration.py's `from models import ...` is also
  one-way now that models no longer re-imports database for Base.

Verification:
- Full backend suite: 2942 passed, 0 failed (vs 2945 pre-refactor;
  the 3-test delta is test-file drift between worktree base and the
  container's existing code, not new failures from wlvxh).
- Import-isolation smoke test: every touched module loads without
  circular errors. database.Base is db_base.Base is models.Base is
  export_models.Base — all point at the same declarative_base.
- grep confirms zero remaining `from cloud_storage.base import`
  references outside the types.py historical-reference docstring.

Bead: enhancedchannelmanager-wlvxh
Unblocks: 0i2vt.4 (SyncTarget models), skqln.2 (session_telemetry).
