+++ backend/cloud_storage/__init__.py
@@ -2,11 +2,11 @@
 Cloud storage adapter framework for export distribution.
 Supports S3, Google Drive, OneDrive, and Dropbox.
 """
-from cloud_storage.base import (
+from cloud_storage.factory import get_adapter
+from cloud_storage.types import (
     CloudStorageAdapter,
-    UploadResult,
     ConnectionTestResult,
-    get_adapter,
+    UploadResult,
 )
 
 __all__ = [
