+++ backend/cloud_storage/dropbox_adapter.py
@@ -5,7 +5,7 @@ import logging
 import time
 from pathlib import Path
 
-from cloud_storage.base import CloudStorageAdapter, UploadResult, ConnectionTestResult
+from cloud_storage.types import CloudStorageAdapter, UploadResult, ConnectionTestResult
 
 logger = logging.getLogger(__name__)
