+++ backend/cloud_storage/s3_adapter.py
@@ -6,7 +6,7 @@ import logging
 import time
 from pathlib import Path
 
-from cloud_storage.base import CloudStorageAdapter, UploadResult, ConnectionTestResult
+from cloud_storage.types import CloudStorageAdapter, UploadResult, ConnectionTestResult
 
 logger = logging.getLogger(__name__)
